pyproject.toml
[project]
name = "geolocation-domain-shift"
version = "0.1.0"
description = "Finetuning VLMs for Geolocation under Domain Shift"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "huggingface-hub>=0.20.0",
    "Pillow>=9.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "tqdm>=4.65.0",
    "open_clip_torch>=2.20.0",
    "matplotlib>=3.7.0",
    "scikit-learn>=1.3.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
